fix to make setup.py get_version more robust
Now it only considers a git repo in which setup.py lives at the top level.
This fixed the problem where running setup.py with a relative path from another
directory which happened to be a git repo, or installing the sdist from inside
some other git repo, would make get_version report the version of the *other*
repo.  This was initially encountered in the readthedocs.org build environment,
where the builds seem to be run within the git repo of the RTD source itself.