Fixed run_kasim to use new syntax for influence maps
In the previous version, KaSim would generate the influence map if a -im
argument was provided at the command line. In version 3, a %def declaration
is added to the kappa file. I modified the run_kasim function to add
the necessary syntax. I left a placeholder for putting in perturbation
language stuff too, but I haven't really fully implemented that yet.