Removed __eq__ methods for components, and tests
Following our discussion about the issues related to the use of the
__eq__ operator for comparison of components, the methods have been
removed. For the time being, Implementation of component comparison will
be made on a model/project basis until a better understanding of
model merging use cases becomes clear.

Tests in test_core.py that had been added in the merge have also been
removed.

Note however, that one test in test_core.py that was added in the merge,
test_compartment_initial_error has been retained; this test insures that
MonomerPatterns specified for an initial condition that are the same in
every aspect (sites, states) except for their compartment, are not
considered to be identical. Similarly, the line added to the
is_equivalent_to method of ComplexPattern that checks for the identity
of the Compartment associated with the ComplexPattern has also been
retained. Both of these additions were introduced in commit
75c1e42acc3afe828fd5539c5f5225392f0395c4.