builder.parameter updates estimate_params based on prior
After Jeremy's removal of the dependency on BayesSB (commit
7ac1d162ee6ba9cf8a3), the specification of whether to estimate
parameters in the builder is no longer indicated by the
estimate=True/False keyword argument, but rather by the "prior" kwarg.
This commit completes this change by populating the estimate_params
array based on whether the value for the prior kwarg is None or has been
given some value (presumably an instance of a bayessb.priors prior).