update travis config and fix some test failures
In the travis config, drop explicit python-numpy from apt-get install list as
we want the version pre-installed in python site-packages. Installing
python-scipy will pull in the numpy .deb anyway, but the one in site-packages
will override it.

Fixed a doctest in integrate.py to reflect a certain numpy exception's new
message which was updated around v1.9. Technically this means that test will
now fail on older numpy versions.

In hello_pysb.py the imports required for simulation and plotting were moved
from the top level to the __main__ block, meaning the model can be imported
without scipy or matplotlib installed (minor but good practice).