Add BNGFileInterface, in addition to BNGConsole
Both BNGFileInterface and BNGConsole have very similar API and usage.
The difference is that BNGConsole executes commands instantly and
waits for the response using the pexpect library. BNGFileInterface
buffers all BNG commands and outputs them to the .bngl file for
execution in one go. BNGFileInterface is typically faster (no
repeated command/response lag) and cross-platform; BNGConsole is
great for interactive use and development.