Another reactions_bidirectional fix when initials and rule orders conflict
Under specific conditions where initial conditions are defined in
a different order to how they appear in rules (see new
test_bidrectional_rules_collapse unit test), the
reaction_cache was not generating hits to combine the reactions
properly. This commit fixes this by checking the cache for a
forward match to (reactants, products) in the cache, as well as
matching the reverse key (products, reactants).