Fix species name fixup in matlab exporter
This wasn't updated when our internal species symbols gained the "__"
prefix, so the prefixes were left out of the regex that matches species
terms. The result was terms like "__y(0)" in the resulting matlab code
which is both incorrect (should be "y(0)") and also invalid matlab syntax
(no leading underscores on symbol names).