Update .travis.yml to build KaSim (squashed cmts)
Add KaSim to path (no code to handle KAPPA_HOME var)

A better way (at least a way more consistent with how the BNG
installation is handled) would be to set a KAPPA_HOME environment
variable and use it to look for the KAPPA install in the same fashion as
in bng.py. For the time being we add KaSim to the path and see if the
tests pass on Travis.

Fix KaSim PATH statement in .travis

KaSim is the executable, not the subdirectory

Get Kappa from GitHub

Install ocamlfind via opam

Fix reversible rule in test_kappa (for Python 3)

Make Kappa with "make all"

Fixed reversible rule in model string