Skip BNGL importer unit tests if .bngl file missing (BNG packaging error)
It turns out BNG 2.2.6-stable releases are missing some .bngl files in the
Linux 64 tarball that are present in the OSX 64 tarball. This was causing
Travis unit tests to fail. An example is Validate/egfr_net.bngl, but there
are several others. I've notified the BNG developers.

For now, just warn the user the file is missing and skip that unit test.
This should be re-addressed when the next BNG release is out.