Handle unicode Observable and Expression names in python 2 (#245)
Due to implementation details in numpy's recarray type, `unicode` objects in Observable and Expression names trigger a numpy exception in the simulator code. This is true even if the names are completely ascii-safe; only the actual string object type matters. This commit converts unicode Observable and Expression names to `str` objects before creating the recarray in the simulator, unless running with Python 3 or a potential future version of numpy that fixes this issue on Python 2. Actual non-ascii characters will still trigger an exception, but the message now tells a modeler how to correct their model.