Fix `is_equivalent_to` for compartmental models
Currently, the `is_equivalent_to` method does not consider species
with a defined Monomer compartment, species compartment or both to
be equivalent, i.e. the following are all different:

 * A() ** C1 # monomer compartment
 * ComplexPattern([A()], compartment=C1) # species compartment
 * ComplexPattern([A() ** C1, compartment=C1) # species and monomer
                                              # compartments

These should all be equivalent, which is fixed in this patch.

I've also added a check in the `is_equivalent_to` method to check
that the ComplexPatterns are concrete, i.e. species, and not
abstract patterns, as the matching rules are subtly different.

Resolves: pysb/pysb#257