Allow rule rate params/exprs with sympy fn names
Defining a parameter or expression using a sympy function name (e.g.
`deg`) will cause the following error:

    SympifyError: SympifyError: <function deg at 0x10f0d0140>

The error is down to `_parse_reaction` in bng.py, which runs `sympy.S(x)`
where `x` is a string such as `'deg'`. Converting the string to the relevant
parameter or expression object resolves this.

Resolves: pysb/pysb#259