Fix PysbFlatExporter with compartmental models (#254)
* Fix PysbFlatExporter with compartmental models
* Fix `is_equivalent_to` for compartmental models

## Original commit messages

* Fix PysbFlatExporter with compartmental models

The PysbFlatExporter fails with compartmental models due to `Compartment`'s
`repr()` implementation, which means parameters and parent compartments
get printed in full rather than just their name, which leads to a
`ComponentDuplicateNameError` with compartmental models when using the
PysbFlatExporter. This commit fixes the issue.

Resolves: pysb/pysb#210

* Fix `is_equivalent_to` for compartmental models

Currently, the `is_equivalent_to` method does not consider species
with a defined Monomer compartment, species compartment or both to
be equivalent, i.e. the following are all different:

 * A() ** C1 # monomer compartment
 * ComplexPattern([A()], compartment=C1) # species compartment
 * ComplexPattern([A() ** C1, compartment=C1) # species and monomer
                                              # compartments

These should all be equivalent, which is fixed in this patch.

I've also added a check in the `is_equivalent_to` method to check
that the ComplexPatterns are concrete, i.e. species, and not
abstract patterns, as the matching rules are subtly different.

Resolves: pysb/pysb#257

* Add error to prevent compartment override with **

In the following scenario with two compartments C1 and C2:

    In [95]: (A() ** C1) ** C2
    Out[95]: A() ** C2

This syntax should throw an error, rather than override the existing
compartment.

See also: pysb/pysb#258

* Deprecate is_equivalent_to with non-concrete pats

* Fix synthesis initials with compartmental models

One initial condition based on __source should be generated for each
compartment. A unit test has been added to check for this.