Kappa tests now in tmpdir, fix unclosed file
The newer versions of kappa seem to produce files in the current
working directory, which leaves cruft after running unit tests.
I've set Kappa's working directory to use the same temp dir as the input
files.

Also, a file handle was opened in _parse_kasim_outfile but not closed.
This has been fixed.