Use mass-action rates in StochKit where possible
Non mass-action rates in StochKit require compilation as C code before
simulation can take place. For large models this takes a long time.

I've added a check to see if a reaction is mass-action, and to export it
as such in StochKit XML if so. This should speed up compilation time.

I've also added an EXTENDED_DEBUG support to StochKitSimulator, which logs
the resultant StochKit XML within PySB.