Save and load SimulationResults to HDF5 files
HDF5 is a hierarchical, binary storage format well suited to storing
matrix-like data. This implementation uses the h5py package.

The file format stores not only the simulation trajectories,
but also metadata relevant to archiving, such as the model itself,
simulation settings, parameter and initial condition values,
simulation date/time, and PySB version.

The goals are to aid reproducibility and to facilitate exchange
of simulation data.