Fix StochKitSimulator for models w/ expressions (#277)
* Fix StochKitSimulator for models w/ expressions

* Speed up StochKit export with loop optimisation

* More StochKit speedy speedy

* Fix error with StochKit error message (!)

* Add expand_obs for symbolic Observable expansion

* Fix tspan length for StochKit

* Fix StochKit doctest

* Use mass-action rates in StochKit where possible

Non mass-action rates in StochKit require compilation as C code before
simulation can take place. For large models this takes a long time.

I've added a check to see if a reaction is mass-action, and to export it
as such in StochKit XML if so. This should speed up compilation time.

I've also added an EXTENDED_DEBUG support to StochKitSimulator, which logs
the resultant StochKit XML within PySB.

* Allow models with 'vol' param with StochKit

* Improve formatting of dict comprehension