Fix link rot in appveyor config (#297)
* Fix broken Dropbox links in appveyor.yml

Public Dropbox links were recently discontinued, so the download links
for pygraphviz on appveyor were broken. I've replaced them with
manually shared Dropbox links.

* Update BioNetGen download URL for appveyor

The link we were using now points to BioNetGen 2.3.0, not our current
dependency of BioNetGen 2.2.6. I've updated the download URL to a
(hopefully stable) link on the Faeder lab website

* Install conda==4.3.21 to avoid conda/weave issue

weave was failing on latest conda (4.3.27) due to a
[build issue with conda](https://stackoverflow.com/questions/46450912/unable-to-execute-x86-64-conda-cos6-linux-gnu-gcc-no-such-file-or-directory)
We will use an earlier version of conda on Travis for now until the
issue is fixed upstream.