Save and load SimulationResults to HDF5 files (#292)
HDF5 is a hierarchical, binary storage format well suited to storing
matrix-like data. This implementation uses the h5py package.

The file format stores not only the simulation trajectories,
but also metadata relevant to archiving, such as the model itself,
simulation settings, parameter and initial condition values,
simulation date/time, and PySB version. A new `custom_attrs` dict
of annotations can be added to a SimulationResult object, the
contents of which will be stored in the HDF5 file as group attributes.

The goals of this feature are to aid reproducibility and to facilitate
exchange of simulation data.