Add `|` reversible rule operator
PySB currently uses `<>` in reversible rules, but this syntax is not
compatible with Python 3, which removed that operator. The current
workaround is to use `!=`, which is suboptimal.

Following discussions with the PySB core team, I have implemented
`|` as a reversible rule operator, and added a
`PendingDeprecationWarning` to `<>`. The `<>` syntax should become
deprecated then removed at a later date.

Old syntax:

    A(b=None) + A(b=None) <> A(b=1) % A(b=1)

New syntax:

    A(b=None) + A(b=None) | A(b=1) % A(b=1)