Add warning if using ScipyOdeSimulator w/o weave
I'm finding that many users of PySB are not aware they can install the
additional "weave" package for a significant speed boost to the
ScipyOdeSimulator. This PR adds a warning if pure Python (lambdify)
mode is being used, to instruct users they may wish to install
weave or use the experimental theano mode.