Add `|` reversible rule operator (#309)
* Fix Kappa interface error display on Python3

KaSim/KaSa interface errors do not display when running PySB under
Python 3 (e.g. #300) due to a missing bytes-to-string conversion.
This PR resolves this using `.decode('utf8')` and adds a unit test
to check errors are raised properly with KaSim.

* Add `|` reversible rule operator

PySB currently uses `<>` in reversible rules, but this syntax is not
compatible with Python 3, which removed that operator. The current
workaround is to use `!=`, which is suboptimal.

Following discussions with the PySB core team, I have implemented
`|` as a reversible rule operator, and added a
`PendingDeprecationWarning` to `<>`. The `<>` syntax should become
deprecated then removed at a later date.

Old syntax:

    A(b=None) + A(b=None) <> A(b=1) % A(b=1)

New syntax:

    A(b=None) + A(b=None) | A(b=1) % A(b=1)

* Update pattern matcher with | for reversible rules