Kappa 4 support, Kappa expression improvements
This PR introduces support for Kappa 4, which is enabled by default.
To use Kappa 3 syntax, use the `pre4syntax=False` argument to
`run_simulation` or `run_static_analysis`. Note that Kappa now
requires agents on the reactant side to correspond to agents on the
product side of a rule (i.e. order is preserved). The "dot" agent
is implemented in PySB as `KappaDot`, and is necessary in
[certain situations documented in the Kappa Manual](https://github.com/Kappa-Dev/KaSim/blob/master/CHANGES.md#agent-order-in-rules-must-be-maintained).

For example, the following Kappa rule:

    // Agent A spawns a copy of agent B
    A(),. -> A(),B() @ 'r'

would be implemented in PySB as:

    Rule('spawnB', A() + KappaDot() >> A() + B(), r)

This PR also improves Expression support for Kappa, implementing
`log`, `exp`, `min`, `max`, `sin`, `cos`, `tan`, `mod` and `sqrt`
functions, if expressions via sympy's `Piecewise`, and the `pi`
symbol.

Fixes: #333
Fixes: #288