networkx/pydot support for reading .dot files
This PR makes pygraphviz support optional and introduces pydot as an
alternative for reading .dot (graphviz) files. pygraphviz support on
Windows on recent versions of Python is patchy at best. pydot is
easier to install and provides an alternative.

When dot file reading is required (kappa interface), the results are
now returned in the (more flexible and mainted) networkx format.
The dot file itself can be read using pydot or pygraphviz (both are
tried by default) - see new pysb.util.read_dot() function.

This change also allows us to move appveyor onto Python 3.6 -
pygraphviz was the hurdle that was previously preventing this.