Fix cupsoda simulator error on Tyson model (#332)
CupSodaSimulator was erroring on the models with rates consisting
of a single Parameter or Expression. Calling .args on these returns
a string, not a pysb.Component object, which fails when we try to
multiply together to get the rate. This commit introduces a check
and workaround in the CupSodaSimulator.

I've also removed the pycuda.autoinit import, which triggers GPU
initialization when pysb.simulator is imported, even if a
CupSodaSimulator is not instantiated. Instead, pycuda is initialized
on demand within the n_blocks function.