Fix Kappa 4 multiple agent synthesis
In Kappa 4, a Ghost agent (placeholder 'dot') must be used to make
the two sides of a reaction rule match. For example, the following
two agent synthesis rule:

    Rule('synthesize_A_and_B', M() + None + None >> M() + A() + B(), k)

This was previously failing (one None was silently dropped). This
has been fixed by modifying ReactionPattern.__add__ and implementing
__radd__. The only caveat is with multiple leading Nones, which will
need to be in parentheses to avoid a TypeError from adding two Nones:

    None + (None + M())

This should also be avoidable by reordering the agents to put the
Nones last, as in the original example.

Fixes: #338