Validate site and state names (#352)
Site names, state names and state values are now validated against the
regex '[_a-z][_a-z0-9]*\Z'.

Imports from BNG or SBML with numeric state values are patched with
leading underscores (a log message is generated per monomer when this
occurs).

Export to Kappa with invalid state values '_*[0-9]*$' are patched
with a leading '__state' (a log message is generated per monomer when
this occurs). This requires the Kappa format_* functions have access
to the model to identify which state values have been changed, so
they're now methods of KappaGenerator. I've left one line wrappers for
each of the original top level functions to maintain API compatibility.