Fix Simulator initials_dict when set by array
The `Simulator` class method `initials_dict` returns initial conditions
as a dictionary, mostly for use by network-free simulators like NFsim.
When initial conditions are set by an array or by a list, species which
were not explicitly set using `Initial()` in the model are ignored.

This would only affect users of the `BngSimulator`, which uses this
method, or anyone who called this method directly after setting
initials with an array or list.

This PR fixes the issue by returning a dictionary with values from the
array. This requires a species list for the model, so a `ValueError`
is raised where this is not available/`generate_equations` hasn't been
called yet.