For the GPU SSA implementation, I found a bug where it would save the state of the system too early. It is supposed to calculate a reaction and a time step. If the time step + time > the current time, the output matrix should be filled with the state up until step + time.
I fixed that here. In doing so, i realize the single step method, which calls each individual step on the gpu, isn't going to work. It would require passing back and forth between gpu and cpu multiple states. It is already a factor of 10 slower than the run_all method (runs entire simulation on gpu). Removing that code.

Also, cuda9 on ubuntu installs into /usr/bin, so I added it to the search nvcc path.