Fix invalid component handling (#431)
Previously, if one added an invalid parameter, rule etc. that generated
an exception in `__init__`, (e.g. a validation error), the component
was still added to the model. In interactive mode, this is annoying, as
the model has to be re-instantiated. Example:

    Model()
    Parameter('a', 'invalid_value')  # raises ValueError
    assert len(model.parameters) == 0  # fixed in this PR
    Parameter('a', 1.0)  # should succeed

This PR fixes this situation by adding the Component to the Model
(by calling `Component.__init__()`) last instead of first.