ScipyOdeSimulator multiprocessing capability (#395)
This PR uses the `concurrent.futures` module to add parallel
simulation capability to the ScipyOdeSimulator. The new
`num_processors` argument to `ScipyOdeSimulator.run()`
can be used to specify the number of processes (defaults to 1).

Since the state sent to processes has to be picklable, some refactoring
was needed:

* New `_get_rhs` function returns an appropriate ODE integration/
Jacobian step function
* The integration code is now in `_integrator_process` function
* The `_integrator_process` function is called by a `SerialExecutor`
if one process is used, otherwise by a `ProcessPoolExecutor`.
This architecture reduces code duplication, since the execution
flow is the same for serial and parallel processing modes.