Fix BngSimulator when supplying initials as params
The Simulator class allows users to change initials by supplying
new underlying parameter values (if both initials and params are
supplied, initials take precedence).

This behaviour was broken in the BngSimulator, which uses a dict
for initials instead of arrays, as some its simulators (e.g. NFsim)
are network free and thus may not have an exhaustive species list.

This PR fixes the issue by fixing the initials_dict implementation,
and also refactors the initials array code to use this code as well,
helping to reduce code duplication.