Fix eqn_mode='python' with local fxns
Fixes an issue when a model with local functions is used
with the ScipyOdeSimulator eqn_mode='python'. Lambdify would
show a NameError and wouldn't pick up model parameters, due to
those not being referenced when local functions are read back
in from BNG during network generation.

Includes a unit test as an example.