Fix observable trajectories on HDF5 load
When loading a SimulationResult from an HDF5 file using the .load()
method, observables are left as zeroes (if the observables were not
saved with the results using the include_obs_exprs=True argument).

The issue is due to the fact that Observable species and coefficients
are not saved in the JSON representation, despite the
include_netgen=True argument. This was missed due to a typo in the
SimulationResult unit tests, also fixed here.

This PR addresses the issue by calculating the Observable species
and coefficients using PySB's built-in pattern matching during
the JSON import process. This approach allows existing HDF5 files
to be used without resaving.