Call BNG using asyncio to avoid deadlock
In rare situations, when BNG outputs a lot of data to STDOUT
quickly and fills the buffer, PySB could lockup due to how
subprocess handles reading pipe buffers. The only way to
avoid this would be to disable streaming output from BNG,
which is often useful for debugging.

This PR calls BNG using Python's asyncio framework.
Reading STDOUT uses a timeout to avoid deadlocking.
If a timeout occurs, the process is checked to see if
it has exited, and any remaining data in the STDOUT
buffer is processed.