Fix repr() of single-monomer complexes w/ compartments  (#550)
* Fix repr() of single-monomer complexes w/ compartments

In BNG, a compartment applied to a complex is distinct from one
applied to a monomer, by use of prefix and postfix notation. In
PySB, one can use parentheses to add a compartment to a complex,
but this fails for a complex of length 1. One can use
`as_complex_pattern()` to explicitly upgrade to a complex from
a monomer, but the string representation and PySB flat exporter
currently fail to respect this.

    A() ** C  # Monomer compartment
    as_complex_pattern(A()) ** C  # Complex compartment

This PR adds `as_complex_pattern()` around complexes with a
single monomer pattern that also have a compartment, to make
sure their string representation and PySB flat export works
as expected.

Co-authored-by: Jeremy Muhlich <jmuhlich@bitflood.org>