Work around Python 3.12 OrderedDict repr in doctest
Python 3.12 changed the OrderedDict repr to emit a dict literal as the
constructor argument instead of a list of tuples. We have one doctest
that returns an OrderedDict, failing on 3.12 without this workaround.
This change adds a temporary monkeypatch to collections.OrderedDict
in the nose module-level fixture, but I don't like it.